third_party/WebKit/LayoutTests/accessibility/inline-text-input-expected.txt

Demonstrates that accessible inline text boxes can be found within an input type=text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axInlineTextBox.role is 'AXRole: AXInlineTextBox'
PASS axInlineTextBox.name is 'Value'
PASS successfullyParsed is true

TEST COMPLETE
